+++ functions/private.ps1
@@ -9,7 +9,7 @@ Function resolvedb {
     $path = $executioncontext.sessionstate.path.GetUnresolvedProviderPathFromPSPath($path)
     [pscustomobject]@{
         Path   = $path
-        Exists = Test-Path -path $path
+        Exists = Test-Path -Path $path
     }
     Write-Verbose "ResolveDB Resolved to $Path"
 }
@@ -26,19 +26,19 @@ Function opendb {
 
 
 Function closedb {
-[cmdletbinding()]
-Param(
-    [System.Data.SQLite.SQLiteConnection]$connection,
-    [System.Data.SQLite.SQLiteCommand]$cmd
-)
-            if ($connection.state -eq 'Open') {
-                Write-Verbose "CloseDB Closing database connection"
-                if ($cmd) {
-                    $cmd.Dispose()
-                }
-                $connection.close()
-                $connection.Dispose()
+    [cmdletbinding()]
+    Param(
+        [System.Data.SQLite.SQLiteConnection]$connection,
+        [System.Data.SQLite.SQLiteCommand]$cmd
+    )
+    if ($connection.state -eq 'Open') {
+        Write-Verbose "CloseDB Closing database connection"
+        if ($cmd) {
+            $cmd.Dispose()
         }
+        $connection.close()
+        $connection.Dispose()
+    }
 }
 Function buildquery {
     [cmdletbinding()]
@@ -56,13 +56,13 @@ Function buildquery {
     Process {
         $names = $InputObject.psobject.Properties.name -join ","
 
-        $inputobject.psobject.Properties | ForEach-Object -Begin { $arr = @()} -process {
+        $inputobject.psobject.Properties | ForEach-Object -Begin { $arr = @() } -Process {
             if ($_.TypeNameofValue -match "String|Int\d{2}|Double|Datetime|long") {
                 $arr += @(, $_.Value)
             }
             elseif ($_.TypeNameofValue -match "Boolean") {
                 #turn Boolean into an INT
-                $arr += @(,($_.value -as [int]))
+                $arr += @(, ($_.value -as [int]))
             }
             else {
                 #only create an entry if there is a value
@@ -70,10 +70,10 @@ Function buildquery {
                     Write-Verbose "Creating cliXML for a blob"
                     #create a temporary cliXML file
                     $out = [system.io.path]::GetTempFileName()
-                    $_.value | Export-Clixml -path $out -encoding UTF8
-                    $in = Get-Content -path $out -Encoding UTF8 -ReadCount 0 -raw
+                    $_.value | Export-Clixml -Path $out -Encoding UTF8
+                    $in = Get-Content -Path $out -Encoding UTF8 -ReadCount 0 -Raw
                     $arr += @(, "$($in)")
-                    remove-item -path $out
+                    Remove-Item -Path $out
                 }
                 else {
                     $arr += ""
@@ -93,14 +93,13 @@ Function buildquery {
 
 } #close buildquery
 
-
 Function frombytes {
     [cmdletbinding()]
     Param([byte[]]$Bytes)
 
     Write-Verbose "Converting from bytes to object"
     $tmpFile = [system.io.path]::GetTempFileName()
-    [text.encoding]::UTF8.getstring($bytes) | Out-file -FilePath $tmpfile -Encoding utf8
+    [text.encoding]::UTF8.getstring($bytes) | Out-File -FilePath $tmpfile -Encoding utf8
     Import-Clixml -Path $tmpFile
     if (Test-Path $tmpfile) {
         Remove-Item $tmpFile