+++ functions/public.ps1
@@ -12,15 +12,14 @@ Function Open-MySQLiteDB {
         [string]$Path
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
-
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
     } #begin
 
     Process {
         $db = resolvedb $Path
         if ($db.exists) {
             if ($pscmdlet.shouldprocess($db.path)) {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Opening connection to $Path "
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Opening $Path"
                 opendb $db.path
             }
         }
@@ -28,8 +27,7 @@ Function Open-MySQLiteDB {
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
-
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 
 } #close Open-MySQLiteDB
@@ -45,12 +43,12 @@ Function Close-MySQLiteDB {
         [switch]$Passthru
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
 
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Closing connection to $($connection.DataSource)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Closing source $($connection.DataSource)"
         if ($pscmdlet.ShouldProcess($Connection.DataSource)) {
             $connection.close()
             if ($passthru) {
@@ -63,8 +61,7 @@ Function Close-MySQLiteDB {
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
-
+        Write-Verbose "[$((Get-Date).TimeOfDay)] ending $($myinvocation.mycommand)"
     } #end
 
 } #close Close-MySQLiteDB
@@ -93,12 +90,12 @@ Function ConvertFrom-MySQLiteDB {
         [switch]$RawObject
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
 
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing database $Path "
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Using path $Path "
         $file = resolvedb -Path $path
         if ($file.exists) {
             $connection = opendb -Path $file.path
@@ -107,11 +104,11 @@ Function ConvertFrom-MySQLiteDB {
             Throw "Failed to find database file $($file.path)"
         }
         #verify table exists
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Verifying table $tablename"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Verify table $tablename"
         $tables = Get-MySQLiteTable -connection $connection -KeepAlive
         if ($tables.name -contains $tablename) {
             $query = "Select * from $tablename"
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Importing from table $Tablename"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Found $Tablename"
             Try {
                 $raw = Invoke-MySQLiteQuery -connection $connection -Query $query -As object -KeepAlive -ErrorAction stop
             }
@@ -122,7 +119,7 @@ Function ConvertFrom-MySQLiteDB {
                 #bail out
                 return
             }
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($raw.count) items"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Found $($raw.count) items"
 
             <#
                 find a mapping table using this priority list
@@ -133,15 +130,15 @@ Function ConvertFrom-MySQLiteDB {
             #>
             switch ($pscmdlet.ParameterSetName) {
                 "hash" {
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] User specified property map"
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] User specified property map"
                     $map = $PropertyMap
                     If ($TypeName) {
                         $oTypename = $TypeName
                     }
                 }
                 "table" {
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting property map from $PropertyTable"
-                    $map = Invoke-MySQLiteQuery -Connection $connection -Query "Select * from $propertytable" -KeepAlive -as Hashtable
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] $PropertyTable"
+                    $map = Invoke-MySQLiteQuery -Connection $connection -Query "Select * from $propertytable"-KeepAlive -as Hashtable
                     if ($typename) {
                         $oTypename = $TypeName
                     }
@@ -152,7 +149,7 @@ Function ConvertFrom-MySQLiteDB {
 
                 }
                 "raw" {
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Writing raw objects to the pipeline"
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] Writing raw objects to the pipeline"
                     $raw
                 }
             }
@@ -161,11 +158,11 @@ Function ConvertFrom-MySQLiteDB {
                 foreach ($item in $raw) {
                     $tmpHash = [ordered]@{}
                     if ($oTypename) {
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Inserting typename $oTypename"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] Adding typename $oTypename"
                         $tmpHash.Add("PSTypename", $oTypename)
                     }
                     foreach ($key in $map.keys) {
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ...$key"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] Adding key $key"
                         $name = $key
                         #if value of the raw object is byte[], assume it is an exported clixml file
                         if ($item.$key.gettype().name -eq 'Byte[]') {
@@ -177,20 +174,20 @@ Function ConvertFrom-MySQLiteDB {
                         $value = $v -as $($($map[$key] -as [type]))
                         $tmpHash.Add($name, $value)
                     } #foreach key
-                    New-Object -typename PSObject -property $tmpHash
+                    New-Object -TypeName PSObject -Property $tmpHash
                 } #foreach item
             } #if $map
         } #if table found
         else {
-            Write-Warning "Failed to find a table called $Tablename in $($file.path)"
+            Write-Warning "Failed to find a table called $Tablename $($file.path)"
         }
 
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
         closedb -connection $connection
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
 
     } #end
 
@@ -206,14 +203,14 @@ Function Get-MySQLiteTable {
         [ValidateNotNullOrEmpty()]
         [alias("database", "fullname")]
         [string]$Path,
-        [Parameter(Position = 0, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
+        [Parameter(Position = 0, ValueFromPipeline, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
         [System.Data.SQLite.SQLiteConnection]$Connection,
         [Parameter(HelpMessage = "Do not close the connection.", ParameterSetName = "connection")]
         [switch]$KeepAlive,
         [switch]$Detail
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
     } #begin
 
     Process {
@@ -222,7 +219,7 @@ Function Get-MySQLiteTable {
         }
         if ($pscmdlet.ParameterSetName -eq 'file') {
             $db = resolvedb -Path $path
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting tables from $($db.Path)"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Using path $($db.Path)"
             if ($db.exists) {
                 $iqParams.Add("Path", $db.path)
                 $source = $db.path
@@ -232,8 +229,8 @@ Function Get-MySQLiteTable {
             }
         } #if file parameter set
         else {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using an existing connection $($connection.ConnectionString)"
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] KeepAlive is $KeepAlive"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Using connection $($connection.ConnectionString)"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] KeepAlive is $KeepAlive"
             $iqParams.Add("Connection", $Connection)
             $iqParams.Add("KeepAlive", $KeepAlive)
             #parse out the path from the connection
@@ -244,10 +241,10 @@ Function Get-MySQLiteTable {
         if ($tablenames) {
 
             if ($Detail) {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting table details"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Getting table details"
                 foreach ($item in $Tablenames.name) {
                     $iqParams.query = "PRAGMA table_info($item)"
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($iqparams.query)"
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] $($iqparams.query)"
                     $details = Invoke-MySQLiteQuery @iqParams
                     foreach ($tbl in $details) {
                         [pscustomobject]@{
@@ -266,7 +263,6 @@ Function Get-MySQLiteTable {
                     Source = $source
                     Name   = $Tablenames.name
                 }
-
             }
         }
         else {
@@ -279,7 +275,7 @@ Function Get-MySQLiteTable {
             closedb -connection $connection
         }
 
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 
 } #close Get-MySQLiteTable
@@ -305,7 +301,7 @@ Function ConvertTo-MySQLiteDB {
         [switch]$Force
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
         $file = resolvedb -Path $path
         if ($Append) {
             if ($file.exists) {
@@ -332,7 +328,7 @@ Function ConvertTo-MySQLiteDB {
                         #bail out
                         return
                     }
-                    Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Created database at $($db.fullname)"
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] Opening database $($db.fullname)"
                     $connection = opendb $db.fullname
                 }
             }
@@ -353,7 +349,7 @@ Function ConvertTo-MySQLiteDB {
         foreach ($object in $Inputobject) {
 
             if ($TableExists) {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding object to the table"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Adding object to the table"
                 $iqParams.query = buildquery -InputObject $object -Tablename $TableName
                 if ($pscmdlet.ShouldProcess("object", "Add to table $Tablename")) {
                     Invoke-MySQLiteQuery @iqParams
@@ -365,9 +361,9 @@ Function ConvertTo-MySQLiteDB {
                 #convert types as necessary. Table types can be Text, Int, Real or Blob
                 if ($PSCmdlet.ShouldProcess("PropertyMap", "Create Table")) {
                     $object.psobject.properties |
-                        foreach-object -begin {
+                    ForEach-Object -Begin {
                         $prop = [ordered]@{}
-                    } -process {
+                    } -Process {
                         $prop.Add($_.Name, "Text")
                     }
 
@@ -377,7 +373,7 @@ Function ConvertTo-MySQLiteDB {
                     else {
                         $name = "propertymap_{0}" -f ($object.psobject.typenames[0].replace(".", "_"))
                     }
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating property map table $name"
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] $name"
                     $prop | Out-String | Write-Verbose
                     $newTblParams = @{
                         Connection       = $Connection
@@ -395,18 +391,18 @@ Function ConvertTo-MySQLiteDB {
                     Invoke-MySQLiteQuery @iqParams
                 }
 
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating property hashtable"
+                Write-Verbose "[$((Get-Date).TimeOfDay)]  Creating property hashtable"
                 #get the property names and types
                 $properties = $object.psobject.properties
                 $thash = [ordered]@{}
                 Foreach ($prop in $properties) {
                     Switch -Regex ($prop.TypeNameofValue) {
-                        "Int32$" {$sqltype = "Int" }
-                        "Int64$" {$sqltype = "Real"}
-                        "^System.Double$" {$sqltype = "Real"}
-                        "^System.DateTime" {$sqltype = "Text"}
-                        "^System.String$" {$sqltype = "Text"}
-                        "^System.Boolean$" {$sqltype = "Int"}
+                        "Int32$" { $sqltype = "Int" }
+                        "Int64$" { $sqltype = "Real" }
+                        "^System.Double$" { $sqltype = "Real" }
+                        "^System.DateTime" { $sqltype = "Text" }
+                        "^System.String$" { $sqltype = "Text" }
+                        "^System.Boolean$" { $sqltype = "Int" }
                         default {
                             $sqltype = "Blob"
                         }
@@ -414,14 +410,14 @@ Function ConvertTo-MySQLiteDB {
                     $thash.Add($prop.Name, $sqltype)
                 } #foreach prop
 
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating table: $Tablename"
                 if ($pscmdlet.ShouldProcess($tablename, "Create table")) {
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] Creating table $Tablename"
                     $newTblParams.ColumnProperties = $thash
                     $newtblParams.tablename = $Tablename
                     New-MySQLiteDBTable @newtblParams
                 }
 
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Inserting the first object into the table"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Inserting the first object into the table"
                 #insert the first object into the new table
                 $iqParams.query = buildquery -InputObject $object -Tablename $TableName
 
@@ -437,7 +433,7 @@ Function ConvertTo-MySQLiteDB {
         if ($connection.State -eq "open") {
             closedb -connection $connection
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 
 } #close Convert-MySQLiteDB
@@ -459,7 +455,7 @@ Function New-MySQLiteDB {
     )
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
 
         $db = resolvedb $Path
 
@@ -469,7 +465,7 @@ Function New-MySQLiteDB {
         }
         else {
             If (($db.exists) -AND $Force) {
-                Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Removing previous database at $($db.path)"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Removing $($db.path)"
                 Remove-Item -Path $db.path
             }
 
@@ -487,9 +483,9 @@ Function New-MySQLiteDB {
         }
     } #begin
     Process {
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing new database: $($db.Path)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($db.Path)"
         if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding Metadata table"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Adding Metadata table"
 
             [string]$query = "CREATE TABLE Metadata (Author TEXT,Created TEXT,Computername TEXT,Comment TEXT);"
 
@@ -501,26 +497,26 @@ Function New-MySQLiteDB {
 
             $query = "Insert Into Metadata (Author,Created,Computername,Comment) Values ('$($meta.author)','$($meta.created)','$($meta.computer)','$($meta.comment)')"
 
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Execute non-query $query"
             if ($pscmdlet.ShouldProcess($query)) {
                 $cmd.CommandText = $query
                 [void]$cmd.ExecuteNonQuery()
             }
         }
         else {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] There is no open database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] There is no open database connection"
         }
     } #process
     End {
         if ($connection.state -eq 'Open') {
-            Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
             $connection.close()
             $connection.Dispose()
         }
         if ($Passthru -AND (Test-Path $db.path)) {
-            Get-Item -path $db.path
+            Get-Item -Path $db.path
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
 
     } #end
 
@@ -565,7 +561,7 @@ Function New-MySQLiteDBTable {
     )
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
     } #begin
     Process {
 
@@ -579,10 +575,10 @@ Function New-MySQLiteDBTable {
             else {
                 Write-Warning "Cannot find the database file $($db.path)."
             }
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing database: $($db.Path)"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Using path $($db.Path)"
         }
         else {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using an existing connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Using an existing connection"
         }
         if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
 
@@ -590,8 +586,8 @@ Function New-MySQLiteDBTable {
             #test if table already exists
             if (-not $Force) {
                 $cmd.CommandText = "SELECT name FROM sqlite_master WHERE type='table' AND name='$Tablename' COLLATE NOCASE;"
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Testing for table $Tablename"
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($cmd.CommandText)"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Using table $Tablename"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] $($cmd.CommandText)"
 
                 $ds = New-Object System.Data.DataSet
                 $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
@@ -616,19 +612,19 @@ Function New-MySQLiteDBTable {
             }
 
             If ($TableFree ) {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding Table: $Tablename"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Creating table $Tablename"
                 [string]$query = "CREATE TABLE $tablename "
 
                 if ($pscmdlet.ParameterSetName -match 'typed') {
                     $keys = $ColumnProperties.Keys
-                    $primary = $keys | Select-Object -first 1`
+                    $primary = $keys | Select-Object -First 1`
                     $primaryType = $ColumnProperties.item($Primary)
-                    $cols = $keys | Select-Object -skip 1
-                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Setting Primary Key: $primary <$PrimaryType>"
+                    $cols = $keys | Select-Object -Skip 1
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] $primary $PrimaryType>"
                     $query += "($Primary $PrimaryType PRIMARY KEY"
                     foreach ($col in $cols) {
                         $colType = $ColumnProperties.item($col)
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding column $col <$($coltype)>"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] $col $($coltype)>"
                         $query += ",$col $coltype"
                     }
                     $query += ");"
@@ -640,22 +636,22 @@ Function New-MySQLiteDBTable {
 
                 $cmd = $connection.CreateCommand()
                 $cmd.CommandText = $query
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] $query"
                 if ($pscmdlet.ShouldProcess($query)) {
                     [void]$cmd.ExecuteNonQuery()
                 }
             }
         }
         else {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] There is no open database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] There is no open database connection"
         }
     } #process
     End {
         if ($connection.state -eq 'Open' -AND (-Not $KeepAlive)) {
-            Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
             closedb -connection $connection -cmd $cmd
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 } #new-MySQLitedbtable
 
@@ -664,13 +660,13 @@ Function Invoke-MySQLiteQuery {
     [alias("iq")]
     [outputtype("None", "PSCustomObject", "System.Data.Datatable", "Hashtable")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ParameterSetName = "file", ValueFromPipelineByPropertyName)]
+        [Parameter(Position = 1, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ParameterSetName = "file", ValueFromPipelineByPropertyName)]
         [Alias("fullname", "database")]
         [ValidateNotNullOrEmpty()]
         [string]$Path,
-        [Parameter(Position = 0, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
+        [Parameter(Position = 1, ValueFromPipeline, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
         [System.Data.SQLite.SQLiteConnection]$Connection,
-        [Parameter(Position = 1, Mandatory, HelpMessage = "Enter a SQL query string")]
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter a SQL query string")]
         [string]$Query,
         [Parameter(HelpMessage = "Keep the connection alive.", ParameterSetName = "connection")]
         [switch]$KeepAlive,
@@ -679,12 +675,12 @@ Function Invoke-MySQLiteQuery {
         [string]$As = "object"
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
 
     } #begin
     Process {
         if ($pscmdlet.ParameterSetName -eq 'file') {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using file $path"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] $path"
             $file = resolvedb -path $path
 
             If ($file.exists) {
@@ -695,11 +691,11 @@ Function Invoke-MySQLiteQuery {
             }
         }
         else {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using existing connection $($connection.ConnectionString)"
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] KeepAlive is $KeepAlive"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Using connecton $($connection.ConnectionString)"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] KeepAlive is $KeepAlive"
         }
 
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Invoke query '$query'"
         if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
 
             $cmd = $connection.CreateCommand()
@@ -711,14 +707,14 @@ Function Invoke-MySQLiteQuery {
                     "^([Ss]elect (\w+|\*)|(@@\w+ AS))|([Pp]ragma \w+)" {
                         #   "^Select (\w+|\*)|(@@\w+ AS)" {
                         if ($As -eq "datatable") {
-                            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Datatable"
+                            Write-Verbose "[$((Get-Date).TimeOfDay)] Datatable output"
                             $ds = New-Object System.Data.DataSet
                             $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
                             [void]$da.fill($ds)
                             $ds.Tables
                         }
                         else {
-                            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ExecuteReader"
+                            Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteReader"
                             $reader = $cmd.executereader()
                             #convert datarows to a custom object
                             while ($reader.read()) {
@@ -744,15 +740,15 @@ Function Invoke-MySQLiteQuery {
                         Break
                     }
                     "@@" {
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ExecuteScalar"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteScalar"
                         [void]$cmd.ExecuteScalar()
                         Break
                     }
                     Default {
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ExecuteNonQuery"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] ExecuteNonQuery"
                         #modify query to use Transactions
                         $Revised = "BEGIN TRANSACTION;$($cmd.CommandText);COMMIT;"
-                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using Transactions: $Revised"
+                        Write-Verbose "[$((Get-Date).TimeOfDay)] $Revised"
                         $cmd.CommandText = $revised
                         [void]$cmd.ExecuteNonQuery()
                         Break
@@ -764,10 +760,10 @@ Function Invoke-MySQLiteQuery {
     End {
         #if the connection was passed as a parameter, do not close it. The generating command is responsible
         if ( (($connection.state -eq 'Open') -AND ($pscmdlet.ParameterSetName -eq 'file')) -OR (($connection.state -eq 'Open') -AND (-Not $KeepAlive)) ) {
-            Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
             closedb -connection $connection -cmd $cmd
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 } #close invoke-MySQLitequery
 
@@ -783,8 +779,7 @@ Function Get-MySQLiteDB {
         [string]$Path
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
-
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
     } #begin
 
     Process {
@@ -792,13 +787,13 @@ Function Get-MySQLiteDB {
         $file = resolvedb -path $path
 
         If ($file.exists) {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting database information from $($file.path)"
-            $thisdb = Get-Item -path $file.path
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Opening $($file.path)"
+            $thisdb = Get-Item -Path $file.path
 
             $connection = opendb -Path $file.path
 
             $tables = Get-MySQLiteTable -Connection $connection -KeepAlive | Select-Object -ExpandProperty Name
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($tables.count) Tables"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Found $($tables.count) Tables"
 
             $pgsize = (Invoke-MySQLiteQuery -connection $connection -query "PRAGMA page_size" -KeepAlive ).page_size
             $pgcount = (Invoke-MySQLiteQuery -connection $connection -query "PRAGMA page_count" -KeepAlive ).page_count
@@ -828,10 +823,10 @@ Function Get-MySQLiteDB {
 
     End {
         if ($connection.open) {
-            Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+            Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
             closedb -connection $connection
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
     } #end
 
 } #close Get-MySQLiteDB